refactor(zkevm): move test path for worst log opcodes
